+++ src/main/java/sky/pro/course3/homeworks/service/StudentService.java
@@ -20,7 +20,7 @@ public class StudentService {
     }
 
     public Student findStudent(long id) {
-        return studentRepository.findById(id).get();
+        return studentRepository.findById(id).orElse(null);
     }
 
     public Student editStudent(Student student) {
@@ -28,7 +28,11 @@ public class StudentService {
     }
 
     public void deleteStudent(long id) {
-        studentRepository.deleteById(id);
+
+        if (studentRepository.existsById(id)){
+            studentRepository.deleteById(id);
+        }
+
     }
 
     public Collection<Student> getStudentByAge(int age) {
